+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.0] - 2021-05-26
+
+Add support for RDS and simplify schedules syntax.
+
+### Added
+
+- Can now start and stop AWS RDS DB Instance !
+
+### Changed
+
+- New simpler syntax to declare schedule, close to the one used for GCP.
+
+### Fixed
+
+- The Cloudwatch Log Group is now created before the lambda.
+
 ## [0.4.0] - 2021-05-26
 
 Cleanup custom role feature, removing the need of a target and adding some documentation.